@@ -12,34 +12,29 @@
     "@fortawesome/free-solid-svg-icons": "^6.4.2",
     "@fortawesome/react-fontawesome": "^0.2.0",
     "@rails/webpacker": "5.4.4",
+    "@react-google-maps/api": "^2.19.3",
     "aws-sdk": "^2.1467.0",
     "babel-loader": "8.0.4",
-    "css-loader": "^5.2.7",
     "dotenv": "^16.3.1",
     "google-map-react": "^2.2.1",
-    "moment": "^2.29.4",
     "query-string": "^8.1.0",
     "react": "^18.2.0",
     "react-addons-shallow-compare": "^15.6.3",
-    "react-dates": "^21.8.0",
-    "react-day-picker": "^8.8.2",
+    "react-datepicker": "^6.9.0",
     "react-dom": "^18.2.0",
     "react-geocode": "^0.2.3",
     "react-google-autocomplete": "^2.7.3",
-    "react-google-maps": "^9.4.5",
     "react-hook-form": "^7.46.2",
     "react-places-autocomplete": "^7.3.0",
     "react-redux": "^8.1.2",
+    "react-responsive-carousel": "^3.2.23",
     "react-router": "^5.2.0",
     "react-router-dom": "^5.2.0",
-    "react-with-direction": "^1.4.0",
-    "react-with-styles": "^5.0.0",
     "redux": "^4.2.1",
     "redux-logger": "^3.0.6",
     "redux-thunk": "^2.4.2",
-    "style-loader": "2.0.0",
-    "webpack": "^4.46.0",
-    "webpack-cli": "^3.3.12"
+    "webpack-cli": "^5.1.4",
+    "webpack-dev-server": "^5.0.4"
   },
   "description": "This README would normally document whatever steps are necessary to get the application up and running.",
   "version": "1.0.0",
@@ -54,8 +49,8 @@
     "postinstall": "webpack"
   },
   "engines": {
-    "node": "12.5.0",
-    "npm": "6.9.0"
+    "node": ">=16.13.2",
+    "npm": ">=8.1.2"
   },
   "keywords": [],
   "author": "",
@@ -63,6 +58,10 @@
   "devDependencies": {
     "@iconify/icons-simple-icons": "^1.2.70",
     "@iconify/react": "^4.1.1",
-    "webpack-dev-server": "^3"
+    "css-loader": "^7.1.1",
+    "sass": "^1.77.2",
+    "sass-loader": "^14.2.1",
+    "style-loader": "^4.0.0",
+    "webpack": "^5.91.0"
   }
 }
