@@ -1,104 +1,104 @@
-import React from 'react';
-import { withRouter } from 'react-router-dom';
-import Autocomplete from "react-google-autocomplete";
-// import PlacesAutocomplete from 'react-places-autocomplete';
-
+import React, { useEffect, useState, useRef } from 'react';
+import { withRouter, useHistory } from 'react-router-dom';
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
-import { faSearch } from '@fortawesome/free-solid-svg-icons';
+import { faLocationDot, faSearch } from '@fortawesome/free-solid-svg-icons';
 
+const Search = (props) => {
+    const inputRef = useRef(); //sets focus on search bar
+    const [mapLocation, setMapLocation] = useState("");
+    const [mapLat, setMapLat] = useState();
+    const [mapLng, setMapLng] = useState();
+    const history = useHistory();
 
-class Search extends React.Component {
-    constructor(props) {
-        super(props)
-        
-    this.state = {
-        mapLocation: "",
-        lat: 37.8651,
-        lng: 119.5383
+    useEffect(() => {
+        inputRef.current.focus();
+    }, [])
+
+    
+    const simulateArrowDown = () => {
+        const event = new KeyboardEvent('keydown', { keyCode: 40 });
+        inputRef.current.dispatchEvent(event);
     };
 
-    this.handleSubmit = this.handleSubmit.bind(this);
-    this.handleInput = this.handleInput.bind(this);
+    const simulateEnter = () => {
+        const event = new KeyboardEvent('keydown', { keyCode: 13 });  
+        inputRef.current.dispatchEvent(event);
     }
 
-    handleInput(e) {
-        this.setState({
-            mapLocation: e.target.value
-        });
-    }
-
-    handleSubmit(e) {
+    const handleInput = (e) => {
         e.preventDefault();
-        this.props.receiveLocation(this.state);
-        // if (this.state.lat == "" || this.state.lng == "") {
-        //     this.setState({
-        //         lat: 37.8651,
-        //         lng: 119.5383
-        //     })
-        // }
-        this.props.history.push({
-            pathname: `/search/${this.state.lat},${this.state.lng}`,
-            state: this.state
+        setMapLocation(e.target.value);
+    }
+
+    const handleSubmit = (e) => {
+        e.preventDefault();
+
+        if (mapLat && mapLng) {
+            const state = {
+                mapLocation,
+                lat: mapLat,
+                lng: mapLng
+            }
+            props.receiveLocation(state);
+            
+            history.push({
+                pathname: `/search/${mapLat},${mapLng}`,
+                state
+            })
+        } else {
+            simulateArrowDown()
+            simulateEnter()
+            simulateEnter()            
+        }
+    }
+
+    const autoCompletePlace = () => {
+        const res = (new google.maps.places.Autocomplete(inputRef.current));
+        res.addListener('place_changed', async () => {
+            const address = await res.getPlace().formatted_address;
+            const place = await res.getPlace();
+            const lat = await place.geometry.location.lat();
+            const lng = await place.geometry.location.lng();
+            const mapRes = address ? address : place.name;
+            setMapLocation(mapRes);
+            setMapLat(lat);
+            setMapLng(lng);
         });
     }
 
-    componentDidMount() {
-        let input = document.getElementById('splash_search');
-        let autocomplete = new google.maps.places.Autocomplete(input);
+    useEffect(() => {
+        autoCompletePlace();
+    }, [])
+    
 
-        let mapLocation;
-        let that = this;
-        autocomplete.addListener('place_changed', () => {
-            let address = autocomplete.getPlace().formatted_address;
-            let place = autocomplete.getPlace();
+    return (
+        <form className="form_search" autocomplete="on">
+            <div className="superSearch">
 
-            let lat = place.geometry.location.lat();
-            let lng = place.geometry.location.lng();
-            mapLocation = address ? address : autocomplete.getPlace().name;
-            that.setState({
-                mapLocation: autocomplete.getPlace().name,
-                lat: lat,
-                lng: lng
-            });
-        });   
-    }
+                <div className="searchBar">
+                    <span className="fasearch">
+                        <FontAwesomeIcon icon={faLocationDot} />
+                    </span>
 
-    render () {
-
-        return (
-            <form className="searchWrapper searchIndex" onSubmit={this.handleSubmit} autocomplete="on">
-                <div className="superSearch">
-
-                    <div className="searchBar">
-                        <span className="fasearch"><FontAwesomeIcon icon={faSearch} /></span>
-                        <input 
+                    <input 
                         id="splash_search" 
-                        className="search" 
+                        className="search"
                         type="search" 
                         placeholder="Start with somewhere like Yosemite Valley!" 
-                        value={this.state.mapLocation}
-                        onChange={this.handleInput}
-                        />
-                    </div>
-                    {/* <div className="datebox"> */}
-                        {/* <div className="filterButton"> */}
-                            {/* future modal */}
-                            {/* <input className="date" type="date" /> */}
-                        {/* </div> */}
-                    {/* </div> */}
+                        ref={inputRef}
+                        onChange={handleInput}
+                        autoComplete="country-name"
+                    />
 
-                    {/* <div className="dropdownMenu"> */}
-                        {/* <select className="menuSelect">
-                            {house.map((tent, i) =>
-                                <option className="menuSelected" key={i} value={i}>{tent}</option>
-                            )}
-                        </select> */}
-                    {/* </div> */}
-                    <button className="searchButton">Search</button>
                 </div>
-            </form>
-        ) 
-    }
+                
+                <button className="searchButton" onClick={handleSubmit}>
+                    <FontAwesomeIcon icon={faSearch} />
+                    <p className='margin-left-5'>Search</p>
+                </button>
+            </div>
+        </form>
+    ) 
 };
 
 export default withRouter(Search);
