@@ -1,43 +1,43 @@
 import React from "react";
-import { withRouter } from "react-router-dom";
-import "react-dates/initialize";
-import "react-dates/lib/css/_datepicker.css";
-import { SingleDatePicker } from "react-dates";
-import moment from 'moment'
-import DefaultTheme from 'react-dates/lib/theme/DefaultTheme';
-import "moment/locale/en-gb"
+import DatePicker from "react-datepicker";
+import "react-datepicker/dist/react-datepicker.css";
 
 class BookingUpdateForm extends React.Component {
   constructor(props) {
     super(props);
 
     this.state = {
-      price: this.props.listings[this.props.booking.listing_id].price,
-      check_in: null,
-      check_out: null,
-      capacity: this.props.listings[this.props.booking.listing_id].capacity,
+      price: this.props.booking.price || "error",
+      check_in: this.props.booking.check_in,
+      check_out: this.props.booking.check_out,
+      capacity: this.props.booking.capacity,
       id: this.props.booking.id,
-      listing_name: this.props.listings[this.props.booking.listing_id].name,
       user_id: this.props.booking.user_id,
-      focusedStart: null,
-      focusedEnd: null,
+      listing_name: this.props.booking.listing_name,
+      max_capacity: "0",
+      listing_id: this.props.booking.listing_id,
+      min_stay: 0,
+      guest: this.props.booking.capacity > 1 ? "Guests" : "Guest"
     };
 
-    this.highlighted = this.highlighted.bind(this);
     this.handleSubmit = this.handleSubmit.bind(this);
     this.handleCapacity = this.handleCapacity.bind(this);
     this.handlePrice = this.handlePrice.bind(this);
+    this.handleCheckIn = this.handleCheckIn.bind(this);
   }
 
   componentDidMount() {
-    this.setState({
-      price: this.props.listings[this.props.booking.listing_id].price,
-    });
+    this.props.fetchListing(this.state.listing_id)
+      .then(res => {
+        this.setState({
+          max_capacity: res.listing.capacity,
+          min_stay: res.minimum_nights
+        })
+      })
   }
 
   handlePrice(date) {
-    const mseconds =
-      Date.parse(date) - Date.parse(this.state.check_in);
+    const mseconds = Date.parse(date) - Date.parse(this.state.check_in);
     const days = (mseconds / (1000 * 60 * 60 * 24)) * this.props.listings[this.props.booking.listing_id].price;
     this.setState({
       price: days,
@@ -49,10 +49,14 @@ class BookingUpdateForm extends React.Component {
     this.setState({
       capacity: event.currentTarget.value,
     });
+    event.currentTarget.value > 1 ? 
+      this.setState({guest: "Guests"}) : this.setState({guest: "Guest"});
   }
 
-  highlighted(day) {
-    return day.isSame(this.state.check_in);
+  handleCheckIn(dates) {
+      const [start, end] = dates;
+      this.setState({ check_in: start, check_out: end });
+      this.handlePrice(end);
   }
 
   handleSubmit(e) {
@@ -65,102 +69,68 @@ class BookingUpdateForm extends React.Component {
       
     } else {
       const booking = {
-        check_in: this.state.check_in.format("DD-MM-YYYY"),
-        check_out: this.state.check_out.format("DD-MM-YYYY"),
+        check_in: this.state.check_in.toISOString(),
+        check_out: this.state.check_out.toISOString(),
         capacity: this.state.capacity,
         price: this.state.price,
         id: this.props.booking.id,
-        listing_name: this.props.listings[this.props.booking.listing_id].name,
+        listing_name: this.state.name,
         user_id: this.props.booking.user_id,
       };
+
       this.props.patchBooking(booking)
-      .then(() => this.props.closeModal())
-      .then(() => this.props.fxn(booking))
+        .then(() => this.props.closeModal())
+        .then(() => this.props.fxn(booking))
     }
   }
 
   render() {
-    let listprice = this.props.listings[this.props.booking.listing_id] ? this.props.listings[this.props.booking.listing_id].price : ""
-    let capacity = this.props.listings[this.props.booking.listing_id] ? this.props.listings[this.props.booking.listing_id].capacity : ""
-
+      const yesterday = new Date(new Date().setDate(new Date().getDate() - 1));
     return (
-      <div className="widget-update-container">
         <div className="widget-container">
-          <form className="wrapper" onSubmit={this.handleSubmit}>
+          <form className="w100" onSubmit={this.handleSubmit}>
             <div className="price-wrapper">
               <div className="price">
-                ${listprice}
+                ${this.state.price}
                 <p>per night</p>
               </div>
             </div>
             <div className="dates-and-guest-content">
-              <div className="col checkin">
-                <div className="label">
-                  Check in
-                    <SingleDatePicker
-                    displayFormat={() => ("MM/DD/YYYY")}
-                    placeholder="Select Start"
-                    date={this.state.check_in} // momentPropTypes.momentObj or null
-                    onDateChange={(date) => this.setState({ check_in: date })} // PropTypes.func.isRequired
-                    focused={this.state.focusedStart} // PropTypes.bool
-                    onFocusChange={({ focused }) => this.setState({ focusedStart: focused })} // PropTypes.func.isRequired
-                    id="start" // PropTypes.string.isRequired,
-                    verticalSpacing={0}
-                    isDayHighlighted={(day) => this.highlighted(day)}
-                    numberOfMonths={1}
-                    daySize={36}
-                    noBorder={true}
-                    hideKeyboardShortcutsPanel={true}
+
+              <div className="booking-label col booking-border margin7">
+                  <p className="flex">Change Dates</p>
+                  <DatePicker
+                      id="datePicker"
+                      showIcon
+                      onChange={this.handleCheckIn}
+                      startDate={this.state.check_in}
+                      endDate={this.state.check_out}
+                      selectsRange
+                      monthsShown={2}
+                      placeholderText="Select Date"
+                      toggleCalendarOnIconClick
+                      excludeDateIntervals={[{ start: "1968/01/01", end: yesterday }]}
                   />
-                </div>
               </div>
-              <div className="col checkout">
-                <div className="label">
-                  Check out
-                    <SingleDatePicker locale="en-gb"
-                    displayFormat={() => ("MM/DD/YYYY")}
-                    placeholder="Select End"
-                    date={this.state.check_out} // momentPropTypes.momentObj or null
-                    onDateChange={(date) => this.setState({ check_out: date })} // PropTypes.func.isRequired
-                    focused={this.state.focusedEnd} // PropTypes.bool
-                    onFocusChange={({ focused }) => this.setState({ focusedEnd: focused })} // PropTypes.func.isRequired
-                    id="end" // PropTypes.string.isRequired,
-                    verticalSpacing={0}
-                    isDayHighlighted={(day) => this.highlighted(day)}
-                    numberOfMonths={1}
-                    daySize={36}
-                    noBorder={true}
-                    hideKeyboardShortcutsPanel={true}
-                    onDateChange={date => this.handlePrice(date)}
-                  />
-                </div>
-              </div>
+
               <div className="col capacity">
                 <div className="label">
-                  Guests
-                    <div className="SingleDatePicker SingleDatePicker_1">
-                    <div className="SingleDatePickerInput SingleDatePickerInput_1">
-                      <div className="DateInput DateInput_1">
-                        <input
-                          type="number"
-                          name="capacity"
-                          className="DateInput_input DateInput_input_1"
-                          placeholder="1"
-                          value={this.state.capacity}
-                          id="capacity_input"
-                          min="1"
-                          max={capacity}
-                          onChange={() => this.handleCapacity}
-                        />
-                      </div>
-                    </div>
-                  </div>
+                  <p>{this.state.guest}</p>
+                  <input
+                    type="number"
+                    name="capacity"
+                    placeholder={this.state.capacity}
+                    id="capacity_input"
+                    min="1"
+                    max={this.state.max_capacity}
+                    onChange={this.handleCapacity}
+                  />
                 </div>
               </div>
+
             </div>
             <button className="booking-button">Update Booking</button>
           </form>
-        </div>
       </div>
     );
   }
