@@ -2,12 +2,11 @@ import BookingUpdateForm from './booking_update_form'
 import { connect } from "react-redux";
 import { closeModal } from "../../actions/modal_actions/modal_actions";
 import { patchBooking } from "../../actions/booking_actions/booking_action";
-import { withRouter } from "react-router-dom";
+import { fetchListing } from "../../actions/listing_actions/listing_actions";
 
 
 const mstp = (state, ownProps) => {
   return ({
-    // currentUser: state.entities.users[state.session.currentUser],
     listings: state.entities.listings,
     bookings: state.entities.bookings,
   });
@@ -15,7 +14,8 @@ const mstp = (state, ownProps) => {
 
 const mdtp = (dispatch) => ({
   closeModal: () => dispatch(closeModal()),
-  patchBooking: (booking) => dispatch(patchBooking(booking))
+  patchBooking: (booking) => dispatch(patchBooking(booking)),
+  fetchListing: (listing) => dispatch(fetchListing(listing))
 });
 
-export default withRouter(connect(mstp, mdtp)(BookingUpdateForm));
+export default connect(mstp, mdtp)(BookingUpdateForm);
