@@ -1,9 +1,6 @@
 import React from 'react';
 import {withRouter} from 'react-router-dom';
 import { Link } from 'react-router-dom';
-// import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
-// import { faGithub, faLinkedin, faAngellist } from '@fortawesome/free-brands-svg-icons';
-
 
 class IndexItem extends React.Component {
     constructor(props) {
@@ -18,17 +15,22 @@ class IndexItem extends React.Component {
     render() {
         return ( 
             <div className="index_container">
+
                 <div className="index_box">
                     <Link to={`/listings/${this.props.listing.id}`}>
-                        <img className="indexPhotos" src={this.props.listing.photoUrls} alt=""/>
+                        <img className="indexPhotos" src={this.props.listing.photoUrls[0]} alt=""/>
                     </Link>
                 </div>
-                <div>
-                    <div className="listing_index_desc">
-                        <div className="list_index_name">{this.props.listing.name}</div>
-                        <div className="listing_index_price">${this.props.listing.price}/day</div>
+
+                <div className="listing_index_desc">
+                    <div className="list_index_name">{this.props.listing.name}</div>
+                    <div className="listing_index_price margin_left2">From 
+                        <div className="bold">
+                            ${this.props.listing.price}
+                        </div>/day
                     </div>
                 </div>
+                
             </div>
         )
     };
