@@ -53,7 +53,7 @@ class Homepage extends React.Component {
         //yosemite
         e.preventDefault();
         this.props.history.push({
-            pathname: `/search/37.8032783,-119.5393594`,
+            pathname: `/search/37.74557009999999,-119.5936038`,
             state: { 
                 lat: 37.8032783,
                 lng: -119.5393594
@@ -84,18 +84,15 @@ class Homepage extends React.Component {
                     < SearchBar />
 
                 <div className="containerBanner">
-                    <img src="https://hipcamp-res.cloudinary.com/image/upload/c_thumb,w_1080/v1498456730/campground-photos/yene977lfve4yhhnokuh.jpg" alt="bannerbar"
+                    <img src="https://hipcamp-res.cloudinary.com/f_auto,c_limit,w_1120,q_auto/campground-photos/gyqfoddi8knyfberfznf.jpg" alt="bannerbar"
                     className="bannerImg"
                     id="bannerImg"
                     />
                     <div className="banner_message">
                         <div >
                                 <span className="banner_font">Explore Dozens of Locations!</span>
-                            {/* <span className="banner_font">Locations!</span> */}
                         </div>
                         <span className="banner_sub_message">Run a search or scroll down below and start exploring!</span>
-                            {/* <span className="banner_sub_message">and start exploring!</span> */}
-
                         <div className="intro_button_container">
                             <button onClick={this.handleInput} className="intro_button">I'm feeling Lucky!</button>
                         </div>
@@ -103,26 +100,52 @@ class Homepage extends React.Component {
                 </div>
 
                 <div className="tilesContainer">
+                    <h1 className="title_listing">Check Out These Areas!</h1>
                     <div className="tiles_tile">
                         <img onClick={this.handleTile1} className="poppingRectangle" src="https://hipcamp-res.cloudinary.com/images/c_thumb,f_auto,q_60,h_953/v1513711999/campground-photos/jxhpqyakfthq93t8on4x/mudita-camel-s-yurt-camels-and-a-yurt-animal-camel.jpg" alt=""/>
                         <img onClick={this.handleTile2} className="poppingRectangle" src="https://hipcamp-res.cloudinary.com/images/c_thumb,f_auto,q_60,h_546/v1494284717/campground-photos/cy19txhitis6xwltkdhu/ardor-wood-farm-glamping-glen-forest-people-dog.jpg" alt=""/>
                         <img onClick={this.handleTile3} className="poppingRectangle" src="https://hipcamp-res.cloudinary.com/images/c_thumb,f_auto,q_60,h_953/v1572899400/campground-photos/hamwjtqwsdqffioglcvq/tiny-house-in-happy-valley-tiny-house-glamping-mountains.jpg" alt=""/>
                     </div>
                 </div>
-                {/* <p>destination getaways.</p> */}
-
+                
                 <div className="vagueLocationsContainer">
+                    <h1 className="title_listing">Our top listings</h1>
                     <div className="vague_tile_list">
-                        <Link to="/listings/46"><img className="vague_picture" src="https://supertramp-mast.s3-us-west-1.amazonaws.com/57056162_10107436356788213_4281326518522609664_o.jpg" alt=""/> </Link>
-                        <Link to="/listings/47"><img className="vague_picture" src="https://supertramp-mast.s3-us-west-1.amazonaws.com/17545579_10105436252223793_1168540811776764446_o.jpg" alt=""/> </Link>
-                        <Link to="/listings/48"><img className="vague_picture" src="https://supertramp-mast.s3-us-west-1.amazonaws.com/55937490_10107418126272313_2116183051628183552_n.jpg" alt=""/> </Link>
+                        <Link to="/listings/174"><img className="vague_picture" src="https://supertramp-mast.s3-us-west-1.amazonaws.com/57056162_10107436356788213_4281326518522609664_o.jpg" alt=""/> </Link>
+                        <Link to="/listings/175"><img className="vague_picture" src="https://supertramp-mast.s3-us-west-1.amazonaws.com/17545579_10105436252223793_1168540811776764446_o.jpg" alt=""/> </Link>
+                        <Link to="/listings/176"><img className="vague_picture" src="https://supertramp-mast.s3-us-west-1.amazonaws.com/55937490_10107418126272313_2116183051628183552_n.jpg" alt=""/> </Link>
                     </div>
 
                     <div className="vague_tile_list">
 
-                        <Link to="/listings/49"><img className="vague_picture" src="https://supertramp-mast.s3-us-west-1.amazonaws.com/17855640_10105438859618553_1862219686291433231_o.jpg" alt=""/> </Link>
-                        <Link to="/listings/53"><img className="vague_picture" src="https://supertramp-mast.s3-us-west-1.amazonaws.com/28828057_10106502756296083_3917095514831156302_o.jpg" alt=""/> </Link>
-                        <Link to="/listings/52"><img className="vague_picture" src="https://supertramp-mast.s3-us-west-1.amazonaws.com/56312133_10107420712439613_2489365651806748672_n.jpg" alt=""/> </Link>
+                        <Link to="/listings/177"><img className="vague_picture" src="https://supertramp-mast.s3-us-west-1.amazonaws.com/17855640_10105438859618553_1862219686291433231_o.jpg" alt=""/> </Link>
+                        <Link to="/listings/204"><img className="vague_picture" src="https://supertramp-mast.s3-us-west-1.amazonaws.com/28828057_10106502756296083_3917095514831156302_o.jpg" alt=""/> </Link>
+                        <Link to="/listings/203"><img className="vague_picture" src="https://supertramp-mast.s3-us-west-1.amazonaws.com/56312133_10107420712439613_2489365651806748672_n.jpg" alt=""/> </Link>
+                    </div>
+                </div>
+
+                <div className=""> 
+                    <h3 className="sp_header">Safety Partners</h3>
+                    <div className='flex'>
+                        <div className='margin-right-15 sp_div green_background flex flex-col'>
+                            <img className='sp_img' src="https://hipcamp-res.cloudinary.com/f_auto,c_limit,w_384,q_auto/internal/recreate_responsibly_text.png"></img>
+                            <a href='https://www.recreateresponsibly.org/' className='flex align_center learn_more'>Learn More</a>
+                        </div>
+
+                        <div className='margin-right-15 sp_div green_background flex flex-col'>
+                            <img className='sp_img' src="https://hipcamp-res.cloudinary.com/f_auto,c_limit,w_384,q_auto/internal/leave_no_trace_text.png"></img>
+                            <a href="https://lnt.org/why/7-principles/" className='flex align_center learn_more'>Learn More</a>
+                        </div>
+
+                        <div className='margin-right-15 sp_div green_background flex flex-col'>
+                            <div className='sp_img'>
+                                <img className='pow' src="https://raw.githubusercontent.com/cowfish813/Supertramp/splash/app/assets/images/protect-our-winters-vector-logo-cropped.svg"></img>
+                                <h1 className="sp_header">Protect Our Winters</h1>
+                                <p className='flex flex-wrap sp_text'>We help passionate outdoor people protect the places and experiences they love from climate change.</p>
+                            </div>
+                            <a href="https://protectourwinters.org/about-pow/" className='flex align_center learn_more'>Learn More</a>
+                        </div>
+
                     </div>
                 </div>
             </div>
