@@ -7,7 +7,9 @@ const configureStore = (preloadedState = {}) =>
     createStore(
         rootReducer, 
         preloadedState, 
-        applyMiddleware(thunk, logger)
+        applyMiddleware(thunk, 
+        // logger
+        )
         ); //add ,logger next to thunk for state/prevstate etc.
 
 export default configureStore
